+++ packages/typespec-rust/CHANGELOG.md
@@ -16,6 +16,7 @@
 ### Other Changes
 
 * Updated to the latest tsp toolset.
+* Client struct fields are now always `pub(crate)`. In addition, internal and helper types are now `pub(crate)` instead of `pub` to help prevent inadvertent exposure.
 
 ## 0.11.0 (2025-03-04)